Implement day choosing and displaying if day is busy
