+++ src/client/timetable/calendar/Calendar.less
@@ -49,41 +49,6 @@
     }
 }
 
-.calendar-date {
-    margin: auto;
-    padding: 7% 5% 5% 5%;
-    line-height: 33px;
-    height: 50px;
-    width: 50px;
-    border: none;
-    background: white;
-    display: inline-block;
-    vertical-align: middle;
-    transition: all 0.3s ease;
-
-    &.chosen-date {
-        background-color: @primary!important;
-        color: #fff !important;
-        border-radius: 100% !important;
-        font-weight: bold !important;
-    }
-
-    &.today {
-        border: red dashed 1px;
-        background-color: white;
-        color: red;
-        border-radius: 100%;
-        font-weight: bold;
-    }
-
-    &:hover {
-        border-radius: 100%;
-        background: #4aa1f3;
-        color: white;
-        text-decoration: none !important;
-    }
-}
-
 .upcoming-events {
     border-radius: 10px;
     hr {
