+++ src/client/timetable/calendar/UpcomingEventsContainer.tsx
@@ -1,37 +1,20 @@
 import React, { useCallback, useMemo } from "react";
 import { Card, Col, Container, Row } from "react-bootstrap";
 import { differenceInCalendarISOWeeks, format, getISODay } from "date-fns";
-import { BaseResponseType } from "../../types";
 import { CourseActivityResponseType } from "../../../types";
-import { requestIsLoaded } from "../../utils";
 import { Loader } from "../../components";
+import {isEmpty} from "lodash";
 
 type Props = {
     chosenDate: Date;
-    activityResponse: BaseResponseType<CourseActivityResponseType>
+    getRelevantActivities: (chosenDate: Date) => CourseActivityResponseType
 };
 
 export const UpcomingEventsContainer: React.FunctionComponent<Props> = ({
     chosenDate,
-    activityResponse
+	getRelevantActivities
 }) => {
-    const relevantActivities = useMemo<CourseActivityResponseType>(() => {
-        if (!requestIsLoaded(activityResponse)) {
-            return {};
-        }
-        return Object.entries(activityResponse.data).reduce((courseActivitiesSoFar: CourseActivityResponseType, [courseCode, activities]) => {
-            return {
-                ...courseActivitiesSoFar,
-                [courseCode]: activities.filter(activity => {
-                   const day = getISODay(chosenDate);
-                   const startDate = new Date(activity.startDate);
-                   const numWeeksFromStartDate = differenceInCalendarISOWeeks(chosenDate, startDate);
-                   return day === activity.dayOfWeek && activity.weeks[numWeeksFromStartDate];
-                })
-            };
-        }, {});
-    }, [activityResponse, chosenDate]);
-
+	const relevantActivities = useMemo(() => getRelevantActivities(chosenDate), [getRelevantActivities, chosenDate]);
     return (
         <Container className="upcoming-events mt-4">
             <Row>
@@ -40,7 +23,7 @@ export const UpcomingEventsContainer: React.FunctionComponent<Props> = ({
                 </header>
                 <hr className="my-2 peach-gradient" />
             </Row>
-            {requestIsLoaded(activityResponse) ?
+            {Object.values(relevantActivities).some((arr) => arr.length !== 0)  ?
                 <Row className="events-container">
                     <Col>
                     {Object.entries(relevantActivities).map(([courseCode, activities]) => (
